Removed all the 'action' stuff I added previously. Removed the validate_action check in ajax.php
Here's why:
The action field in universally being passed as POST data from
the code calling the ajax scripts and then checked against the
same list for all scripts.  There's no binding between an ajax
target and a particular action. Plus, the caller is the one
sending the action. Any attacker could just look at the code
and send the correct action with a request.

A much better approach is to have ajax targets declare allowed
actions, and correlate an action with something the caller
cannot provide itself. e.g 'subnets-add' allows 'edit' and
then check user perms for the 'edit' permission, or something
like that.

The preexisting ones are fine to a called script to determine
what to do, but there's no point as a security feature as implemented