Fixing something bad that happened during a prior merge. Also upstream bug #827
Several files were overwritten with copies from a different module.  Not sure how, but this puts them back.

Also started on (re)fixing #827.  The root problem is that CSRF cookies are being
set named after a particular form/section, but if one opens a form to edit an object (form #1)
then opens another instance of that _same_ form (form #2), even for a different object,
before submitting form #1, the CSRF cookie is overwritten when loading form #2. So
submitting form #1 after form #2 has been opened will *always* fail CSRF validation.

The fix is to name the CSRF cookies after a more unique combination, such as
"formname_objectID_action".  Bear in mind, that opening a form for one object then
opening the exact same form in another tab for the same object and action will _still_
overwrite the CSRF but this is a much narrower case. In short, don't try to do the
same thing to the same object at the same time in two different tabs ;)

This may be an iceberg, the next few commits will probably be fixes for this.